+++ docs/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 |Version|Date|Details|Type|
 |-------|----|-------|----|
-|0.2.1 |31JUL2026|Add dependencies|Patch|
+|0.3.0 |05AUG2026|Add search method and RAG|Minor|
+|0.2.1 |04AUG2026|Add documentation and context|Patch|
+|0.2.1 |02AUG2026|Add dependencies|Patch|
 |0.2.0 |31JUL2026|Add explanation switch|Minor|
 |0.1.0 |30JUL2026|First Package Publish|Minor|
 |0.0.1 |30JUL2026|First push|Patch / Starter|